Fix signed integer comparison issue in read-asdf

+++ examples/read-asdf.cpp
@@ -59,7 +59,7 @@ int main(int argc, char **argv)
     std::cout << alphabet << std::endl;
 
     char *alpha_buff = alphabet.read();
-    for (int i = 0; i < alphabet.get_shape()[0]; i++)
+    for (size_t i = 0; i < alphabet.get_shape()[0]; i++)
     {
         assert(alpha_buff[i] == 'a' + (char) i);
     }
@@ -72,7 +72,7 @@ int main(int argc, char **argv)
     std::uniform_real_distribution<double> dist;
 
     double *rand_buff = random.read();
-    for (int i = 0; i < alphabet.get_shape()[0]; i++)
+    for (size_t i = 0; i < alphabet.get_shape()[0]; i++)
     {
         assert(rand_buff[i] == dist(engine));
     }
